refactor(web): move React hooks to dedicated hooks directory
Move custom React hooks from apps/web/lib/** to apps/web/hooks/**
following a clearer separation of concerns. Hooks are now organized
under feature-scoped subfolders (e.g., hooks/beagle-search/,
hooks/i18n/).

Update all imports and documentation to reflect the new structure.
Non-hook utilities remain in lib/**.

closes #28

Files:
- AGENTS.md
- ARCHITECTURE.md
- apps/web/components/beagle-search/beagle-search-advanced-filters.tsx
- apps/web/components/beagle-search/beagle-search-empty-state.tsx
- apps/web/components/beagle-search/beagle-search-form.tsx
- apps/web/components/beagle-search/beagle-search-page.tsx
- apps/web/components/beagle-search/beagle-search-pagination.tsx
- apps/web/components/beagle-search/beagle-search-results-desktop-table.tsx
- apps/web/components/beagle-search/beagle-search-results-mobile-cards.tsx
- apps/web/components/beagle-search/beagle-search-row-actions.tsx
- apps/web/components/home/__tests__/statistics-section.test.ts
- apps/web/components/home/main-header.tsx
- apps/web/components/home/statistics-section.tsx
- apps/web/components/sidebar/app-header.tsx
- apps/web/components/sidebar/app-sidebar.tsx
- apps/web/hooks/beagle-search/index.ts
- apps/web/hooks/beagle-search/use-beagle-search-ui-state.ts
- apps/web/hooks/i18n/index.ts
- apps/web/hooks/i18n/use-i18n.ts
- apps/web/lib/beagle-search/index.ts
- apps/web/lib/i18n/index.ts
