+++ apps/web/components/beagle-search/beagle-search-pagination.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { beagleTheme } from "@/components/ui/beagle-theme";
 import { BEAGLE_PAGE_SIZE } from "@/lib/beagle-search";
-import { useI18n } from "@/lib/i18n";
+import { useI18n } from "@/hooks/i18n";
 import { cn } from "@/lib/utils";
 
 export function BeagleSearchPagination({
