+++ apps/web/components/sidebar/app-sidebar.tsx
@@ -31,7 +31,8 @@ import {
   useSidebar,
 } from "@/components/ui/sidebar";
 import { cn } from "@/lib/utils";
-import { useI18n, type MessageKey } from "@/lib/i18n";
+import type { MessageKey } from "@/lib/i18n";
+import { useI18n } from "@/hooks/i18n";
 
 type NavItem = {
   labelKey: MessageKey;
