+++ apps/web/hooks/beagle-search/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-beagle-search-ui-state";
