+++ apps/web/hooks/i18n/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-i18n";
