+++ apps/web/lib/beagle-search/index.ts
@@ -1,4 +1,3 @@
 export * from "./birth-year";
 export * from "./legacy-like-match";
 export * from "./types";
-export * from "./use-beagle-search-ui-state";
