+++ apps/web/lib/i18n/index.ts
@@ -1,4 +1,3 @@
 export * from "@/lib/i18n/messages";
 export * from "@/lib/i18n/provider";
 export * from "@/lib/i18n/types";
-export * from "@/lib/i18n/use-i18n";
